PRSNL/backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
uvloop==0.21.0  # High-performance event loop (2-4x async performance boost)
python-dotenv==1.0.0
pydantic-settings==2.1.0
slowapi==0.1.9

# Database
asyncpg==0.29.0
sqlalchemy==2.0.25
pgvector==0.2.0

# Caching & Task Queue
redis==5.0.1
celery==5.3.4
kombu==5.3.4
flower==2.0.1  # Celery monitoring dashboard

# Agent Frameworks
crewai==0.30.11
crewai-tools==0.4.26

# LangChain & LangGraph for workflow orchestration
langgraph>=0.2.0  # State-based AI workflow orchestration
langchain-core>=0.3.0  # Core LangChain abstractions
langchain>=0.3.0  # LangChain framework for prompt templates
langchain-community>=0.3.0  # Community integrations
langchain-openai>=0.2.0  # OpenAI/Azure OpenAI integration

# Web Scraping
beautifulsoup4==4.12.3
httpx==0.26.0
aiohttp==3.9.3
readability-lxml
crawl4ai>=0.3.0

# Image Processing
Pillow==10.2.0
lxml_html_clean
APScheduler==3.11.0

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pytest==7.4.4
pytest-asyncio==0.23.3
pre-commit==3.6.0
bandit==1.7.5
pydocstyle==6.3.0

# Voice Recognition (Offline)
vosk==0.3.44
pywhispercpp>=1.2.0  # whisper.cpp Python bindings

# AI Validation & Quality
guardrails-ai>=0.4.0

# Logging & Monitoring
structlog==24.1.0
sentry-sdk[fastapi]==2.19.2

# Media Processing
yt-dlp
pillow==10.2.0
youtube-transcript-api==0.6.2

# Vision AI
pytesseract==0.3.10
aiofiles==23.2.1
python-multipart==0.0.6

# Monitoring & Observability
prometheus-client==0.19.0
starlette-exporter==0.21.0
python-telegram-bot==20.8
tenacity==8.2.3
psutil==5.9.0

# OpenTelemetry for traces, metrics, and logs (upgraded for guardrails-ai compatibility)
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-asyncpg==0.45b0
opentelemetry-instrumentation-aiohttp-client==0.45b0
opentelemetry-instrumentation-requests==0.45b0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-prometheus==1.12.0rc1
prometheus-fastapi-instrumentator==6.1.0

# AI/ML
openai==1.35.3
scikit-learn==1.3.2
numpy==1.26.2

# RAG & Knowledge Management
haystack-ai>=2.6.0  # Haystack v2 for RAG pipelines
sentence-transformers>=2.2.2  # For local embeddings
qdrant-haystack>=5.0.0  # Qdrant document store integration
pgvector-haystack>=0.4.0  # PGVector integration for Haystack

# AI Quality Monitoring  (will be added later)
# ragas>=0.1.7  # RAG evaluation and monitoring - conflicts with langchain versions

# LangChain & LangGraph for workflow orchestration
langgraph>=0.2.0  # State-based AI workflow orchestration
langchain-core>=0.3.0  # Core LangChain abstractions
langchain>=0.3.0  # LangChain framework for prompt templates
langchain-community>=0.3.0  # Community integrations
langchain-openai>=0.2.0  # OpenAI/Azure OpenAI integration

# Document Processing
markitdown==0.1.2
python-docx==0.8.11
openpyxl==3.1.2
odfpy==1.4.1
python-magic==0.4.27

# Advanced Web Scraping
firecrawl-py>=0.0.14  # Firecrawl API client

# Computer Vision & Image Understanding
open-clip-torch>=2.24.0  # OpenCLIP for image-text matching
torch>=1.11.0  # PyTorch for OpenCLIP
torchvision>=0.12.0  # Vision transformations
ftfy>=6.0.0  # Text normalization for CLIP

# Additional ML Dependencies
nltk>=3.9.1  # Natural language processing
transformers>=4.41.0  # Hugging Face transformers (already included but ensuring version)
spacy>=3.7.0  # Advanced NLP and NER
# Note: Run `python -m spacy download en_core_web_sm` after installation